post-deploy-fixes - added tool default settings

+++ room-stat-service/src/DTO/SignalDTO.php
@@ -20,7 +20,7 @@ class SignalDTO
         return $this->temperature;
     }
 
-    public function humidity(): int
+    public function humidity(): float
     {
         return $this->humidity;
     }
